requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.31.1
python-multipart==0.0.9
faster-whisper==1.0.3
ffmpeg-python==0.2.0
Jinja2==3.1.4
opencc-python-reimplemented==0.1.7
torch==2.4.0
